refactor(db): moved db connection into db folder
