+++ backend/src/db/migrate.ts
@@ -1,7 +1,7 @@
 import * as path from 'path';
 import { promises as fs } from 'fs';
 import { Migrator, FileMigrationProvider } from 'kysely';
-import { db } from '../db';
+import { db } from './db';
 
 async function migrateToLatest() {
   const migrator = new Migrator({
